code review & security update
