+++ client/src/components/Register/Register.js
@@ -34,7 +34,7 @@ class Register extends React.Component {
     })
     .then (response => response.json())
     .then (user => {
-      if(user){
+      if(user.id){
         this.props.loadUser(user);
         this.props.onRouteChange('home');
       }
